+++ Assets/PatchKit Patcher/Editor/PatcherManifestCreator.cs
@@ -8,7 +8,33 @@ namespace PatchKit.Unity.Editor
 {
     public static class PatcherManifestCreator
     {
-        private const int ManifestVersion = 3;
+        private const int ManifestVersion = 4;
+
+        [MenuItem("Tools/PatchKit Patcher/Manifest/Windows")]
+        private static void Test(Manifest manifest)
+        {
+            string targetLocation = EditorUtility.SaveFilePanel("Choose test manifest location", "", "patcher.manifest", "test");
+
+            File.WriteAllText(targetLocation, JsonConvert.SerializeObject(manifest, Formatting.Indented));
+        }
+
+        [MenuItem("Tools/PatchKit Patcher/Manifest/Windows")]
+        private static void TestWindows()
+        {
+            Test(WindowsManifest("BUILD_PATH"));
+        }
+
+        [MenuItem("Tools/PatchKit Patcher/Manifest/Linux")]
+        private static void TestLinux()
+        {
+            Test(LinuxManifest("BUILD_PATH"));
+        }
+
+        [MenuItem("Tools/PatchKit Patcher/Manifest/Osx")]
+        private static void TestOsx()
+        {
+            Test(OsxManifest("BUILD_PATH"));
+        }
 
         [PostProcessBuild, UsedImplicitly]
         private static void PostProcessBuild(BuildTarget buildTarget, string buildPath)
@@ -45,6 +71,13 @@ namespace PatchKit.Unity.Editor
             return new Manifest.Argument { Value = args };
         }
 
+        private static string[] Capabilities()
+        {
+            return new []{
+                "pack1_compression_lzma2",
+            };
+        }
+
         private static Manifest LinuxManifest(string buildPath)
         {
             string patcherExe = Path.GetFileName(buildPath);
@@ -58,6 +91,7 @@ namespace PatchKit.Unity.Editor
 
                 Version = ManifestVersion,
                 Target = "sh",
+                Capabilities = Capabilities(),
                 Arguments = new Manifest.Argument[] {
                     CreateManifestAgument(launchScriptPath),
                     CreateManifestAgument("--exedir={exedir}"),
@@ -79,6 +113,7 @@ namespace PatchKit.Unity.Editor
 
                 Version = ManifestVersion,
                 Target = "{exedir}/" + targetFile,
+                Capabilities = Capabilities(),
                 Arguments = new Manifest.Argument[] {
                     CreateManifestAgument("--installdir", "{installdir}"),
                     CreateManifestAgument("--lockfile", "{lockfile}"),
@@ -97,6 +132,7 @@ namespace PatchKit.Unity.Editor
 
                 Version = ManifestVersion,
                 Target = "open",
+                Capabilities = Capabilities(),
                 Arguments = new Manifest.Argument[] {
                     CreateManifestAgument("{exedir}/" + targetFile),
                     CreateManifestAgument("--args"),